WIP: default logger fixes
Work in progress: created test to ensure constantly asking for the same
default logger does not always (re-) add a handler causing each logged
output to be output by an increasing number of handlers.
Fixed Blueprint.__set_default_logger to not have this problem and also
changed output to stderr from the default of stderr.